+++ CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.32.0] - 2023-02-04 21:41:29
+
+### Added
+
+- Canada net income tree audit.
+
+## [0.31.0] - 2023-02-02 14:47:47
+
+### Added
+
+- OAS params, vars, and tests
+
+## [0.30.0] - 2023-01-31 18:07:39
+
+### Added
+
+- Northern Ontario Energy Credit (NOEC).
+
+## [0.29.0] - 2023-01-31 00:30:36
+
+### Added
+
+- British Columbia family benefit.
+
 ## [0.28.0] - 2023-01-29 15:46:25
 
 ### Added
@@ -259,6 +283,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+[0.32.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.31.0...0.32.0
+[0.31.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.30.0...0.31.0
+[0.30.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.29.0...0.30.0
+[0.29.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.28.0...0.29.0
 [0.28.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.27.1...0.28.0
 [0.27.1]: https://github.com/PolicyEngine/policyengine-canada/compare/0.27.0...0.27.1
 [0.27.0]: https://github.com/PolicyEngine/policyengine-canada/compare/0.26.0...0.27.0